moved the PortfolioDataAccessInterface out of the use_case.signup folder into the more general use_case folder because multiple use case interactors need it.
